# api/openapi.yaml
openapi: 3.0.3
info:
  title: VERAE API
  version: 1.0.0
  description: 'Single source of truth API-contract for frontend and backend integration.

    '
servers:
- url: https://api.verae.local
  description: Example server
security:
- bearerAuth: []
paths:
  /health:
    get:
      tags:
      - Health
      summary: Health check
      description: Simple liveness/readiness check for DevOps and monitoring.
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
      - Auth
      summary: Authenticate user and issue JWT access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyses:
    get:
      tags:
      - Analyses
      summary: List analyses for current user
      description: Returns analyses for authenticated user, sorted by created_at desc. No pagination.
      responses:
        '200':
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAnalysesResponse'
        '401':
          description: Missing/invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Analyses
      summary: Create analysis and enqueue processing job
      description: 'Creates analysis bound to the current authenticated user (`user_id`
        from JWT subject).

        Persists upload metadata and creates processing job with status `queued`.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisRequest'
      responses:
        '202':
          description: Analysis accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAnalysisResponse'
        '401':
          description: Missing/invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyses/{id}:
    get:
      tags:
      - Analyses
      summary: Get analysis processing status
      description: Returns status fields for analysis that belongs to authenticated
        user.
      parameters:
      - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Current analysis state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStatusResponse'
        '401':
          description: Missing/invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Analysis not found for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyses/{id}/result:
    get:
      tags:
      - Analyses
      summary: Get final normalized model result
      description: 'Returns the same shape as POST /v1/risk/predict (iron_index, risk_tier,
        clinical_action, explanations). Only when analysis status is completed.

        '
      parameters:
      - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Final result (same schema as Predict response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictResponse'
        '401':
          description: Missing/invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Analysis not found for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Analysis is not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/risk/predict:
    post:
      tags:
      - Predict
      summary: Predict iron-deficiency risk from manual lab input
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictRequest'
      responses:
        '200':
          description: Prediction response (`ok`) or missing-input response (`needs_input`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AnalysisId:
      name: id
      in: path
      required: true
      description: Analysis UUID
      schema:
        type: string
        format: uuid
  schemas:
    RegisterRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required:
      - access_token
      - token_type
      - expires_in
      - user
      properties:
        access_token:
          type: string
          description: JWT token
        token_type:
          type: string
          enum:
          - Bearer
        expires_in:
          type: integer
          minimum: 1
          description: Token TTL in seconds
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      required:
      - id
      - email
      - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    CreateAnalysisRequest:
      type: object
      required:
      - upload
      - lab
      properties:
        upload:
          $ref: '#/components/schemas/UploadMetadata'
        lab:
          $ref: '#/components/schemas/PredictRequest'
          description: Lab input for model inference (same as POST /v1/risk/predict)
    UploadMetadata:
      type: object
      required:
      - filename
      - content_type
      - size_bytes
      properties:
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          minimum: 1
        checksum_sha256:
          type: string
          description: Hex encoded checksum if available
        source:
          type: string
          description: Optional upload source marker
          example: web
    CreateAnalysisResponse:
      type: object
      required:
      - analysis_id
      - user_id
      - status
      - progress_stage
      - job
      - created_at
      - updated_at
      properties:
        analysis_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          description: Bound from authenticated JWT subject
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        progress_stage:
          $ref: '#/components/schemas/ProgressStage'
        job:
          $ref: '#/components/schemas/JobInfo'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnalysisListItem:
      type: object
      required:
      - analysis_id
      - status
      - created_at
      properties:
        analysis_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - queued
          - processing
          - completed
          - failed
        created_at:
          type: string
          format: date-time
    ListAnalysesResponse:
      type: object
      required:
      - analyses
      properties:
        analyses:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisListItem'
    JobInfo:
      type: object
      required:
      - id
      - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - queued
    AnalysisStatusResponse:
      type: object
      required:
      - analysis_id
      - status
      - progress_stage
      - error_code
      - updated_at
      properties:
        analysis_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        progress_stage:
          $ref: '#/components/schemas/ProgressStage'
        error_code:
          type: string
          nullable: true
          description: Stable machine-readable error code for failed analyses
        updated_at:
          type: string
          format: date-time
    HealthResponse:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - ok
    AnalysisStatus:
      type: string
      enum:
      - queued
      - processing
      - completed
      - failed
    ProgressStage:
      type: string
      enum:
      - queued
      - validating_input
      - preprocessing
      - model_inference
      - postprocessing
      - completed
      - failed
    ErrorResponse:
      type: object
      required:
      - error_code
      - message
      properties:
        error_code:
          type: string
        message:
          type: string
    PredictRequest:
      type: object
      additionalProperties: false
      properties:
        LBXWBCSI:
          type: number
        LBXLYPCT:
          type: number
        LBXMOPCT:
          type: number
        LBXNEPCT:
          type: number
        LBXEOPCT:
          type: number
        LBXBAPCT:
          type: number
        LBXRBCSI:
          type: number
        LBXHGB:
          type: number
        LBXHCT:
          type: number
        LBXMCVSI:
          type: number
        LBXMC:
          type: number
        LBXMCHSI:
          type: number
        LBXRDW:
          type: number
        LBXPLTSI:
          type: number
        LBXMPSI:
          type: number
        LBXSGL:
          type: number
        LBXSCH:
          type: number
        BMXBMI:
          type: number
        BMXHT:
          type: number
        BMXWT:
          type: number
        BMXWAIST:
          type: number
        BP_SYS:
          type: number
        BP_DIA:
          type: number
        RIAGENDR:
          type: integer
        RIDAGEYR:
          type: integer
    PredictResponse:
      oneOf:
      - $ref: '#/components/schemas/PredictResponseOk'
      - $ref: '#/components/schemas/PredictResponseNeedsInput'
    PredictExplanation:
      type: object
      required:
      - feature
      - label
      - impact
      - direction
      - text
      additionalProperties: false
      properties:
        feature:
          type: string
        label:
          type: string
        impact:
          type: number
        direction:
          type: string
          enum:
          - negative
          - positive
        text:
          type: string
    PredictResponseOk:
      type: object
      required:
      - status
      - confidence
      - model_name
      - missing_required_fields
      - iron_index
      - risk_percent
      - risk_tier
      - clinical_action
      - explanations
      additionalProperties: false
      properties:
        status:
          type: string
          enum:
          - ok
        confidence:
          type: string
          enum:
          - low
          - medium
          - high
        model_name:
          type: string
        missing_required_fields:
          type: array
          items:
            type: string
        iron_index:
          type: number
        risk_percent:
          type: number
        risk_tier:
          type: string
          enum:
          - HIGH
          - WARNING
          - GRAY
          - LOW
        clinical_action:
          type: string
        explanations:
          type: array
          items:
            $ref: '#/components/schemas/PredictExplanation'
    PredictResponseNeedsInput:
      type: object
      required:
      - status
      - confidence
      - model_name
      - missing_required_fields
      - iron_index
      - risk_percent
      - risk_tier
      - clinical_action
      - explanations
      additionalProperties: false
      properties:
        status:
          type: string
          enum:
          - needs_input
        confidence:
          type: string
          enum:
          - low
        model_name:
          type: string
        missing_required_fields:
          type: array
          items:
            type: string
          minItems: 1
        iron_index:
          type: number
          nullable: true
        risk_percent:
          type: number
          nullable: true
        risk_tier:
          type: string
          nullable: true
        clinical_action:
          type: string
          nullable: true
        explanations:
          type: array
          items:
            $ref: '#/components/schemas/PredictExplanation'